requirements.txt
google-genai
feedparser
sentence-transformers
python-dotenv
requests